Add background chooser text box
